file_converter/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.2
djangorestframework==3.14.0
docx2txt==0.8
fire==0.5.0
fonttools==4.39.4
kombu==5.3.0
lxml==4.9.2
numpy==1.24.3
opencv-python==4.7.0.72
pandas==2.0.2
pdf2docx==0.5.6
Pillow==9.5.0
prompt-toolkit==3.0.38
PyMuPDF==1.22.3
pypdf2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
termcolor==2.3.0
typing-extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
